// server/types/errors.ts
export enum ErrorCode {
  DECRYPTION_FAILED = 'DECRYPTION_FAILED',
  INVALID_SIGNATURE = 'INVALID_SIGNATURE',
  BLOCKCHAIN_VERIFICATION_FAILED = 'BLOCKCHAIN_VERIFICATION_FAILED',
  DHA_VERIFICATION_FAILED = 'DHA_VERIFICATION_FAILED',
  NPR_VERIFICATION_FAILED = 'NPR_VERIFICATION_FAILED',
  HANIS_VERIFICATION_FAILED = 'HANIS_VERIFICATION_FAILED',
  SAPS_VERIFICATION_FAILED = 'SAPS_VERIFICATION_FAILED',
  QUANTUM_VERIFICATION_FAILED = 'QUANTUM_VERIFICATION_FAILED',
  INVALID_DOCUMENT = 'INVALID_DOCUMENT',
  EXPIRED_DOCUMENT = 'EXPIRED_DOCUMENT',
  TAMPERED_DOCUMENT = 'TAMPERED_DOCUMENT',
  NETWORK_ERROR = 'NETWORK_ERROR',
  ABIS_VERIFICATION_FAILED = 'ABIS_VERIFICATION_FAILED',
  UNKNOWN_ERROR = 'UNKNOWN_ERROR',
  ATTACHMENT_STORAGE_FAILED = 'ATTACHMENT_STORAGE_FAILED',
  PERMIT_ATTACHMENT_FAILED = 'PERMIT_ATTACHMENT_FAILED'
}

export class VerificationError extends Error {
  constructor(
    public code: ErrorCode,
    public message: string,
    public retryable: boolean = false,
    public suggestion?: string
  ) {
    super(message);
    this.name = 'VerificationError';
  }
}